Skip decode logic if `decode: false` is passed
Because it was slightly simpler, we were just always decoding before,
and then we were returning the raw or decoded value depending on what
had been requested. But there may be instances where the decode logic
is failing to handle a value, and it may be beneficial to _not_ run
the decode logic at all…hence this change.